% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_backend.R
\name{check_table_in_db}
\alias{check_table_in_db}
\title{Check for Table in DB
Check if a table exists in the DB}
\usage{
check_table_in_db(connection, table_name)
}
\arguments{
\item{connection}{connection object initalized from get_sql_connection}

\item{table_name}{the name of the table you want to load the book data into}
}
\description{
Check for Table in DB
Check if a table exists in the DB
}
