% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_outputs.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert non standard date formats to standardized dates}
\usage{
convert_to_date(col)
}
\arguments{
\item{col}{date col name}
}
\value{
formated field
}
\description{
Convert non standard date formats to standardized dates
}
