% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_splicing.R
\name{enframe_monthly_analyics}
\alias{enframe_monthly_analyics}
\title{Make the Analytics into a data frame}
\usage{
enframe_monthly_analyics(analyics_repsonse, pluck_value)
}
\arguments{
\item{analyics_repsonse}{Reponse from the analytics API}

\item{pluck_value}{part of the response to grab and enframe}
}
\value{
DF of analytics value you plucked
}
\description{
Make the Analytics into a data frame
}
