% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_outputs.R
\name{extract_bank_accounts_details}
\alias{extract_bank_accounts_details}
\title{Extract Details From the Bank Account JSON}
\usage{
extract_bank_accounts_details(book_info_response)
}
\arguments{
\item{book_info_response}{The response from the get_book_info api call}
}
\value{
list of core_info, activity_info, activing_info_missing, periods
}
\description{
Extract Details From the Bank Account JSON
}
