% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_outputs.R
\name{extract_nested_df}
\alias{extract_nested_df}
\title{Extract Details From the Bank Account JSON}
\usage{
extract_nested_df(df, col_with_data, pluck_value)
}
\arguments{
\item{df}{DF containing the data to extract}

\item{col_with_data}{column containing the data you want to extract}

\item{pluck_value}{value}
}
\value{
list of core_info, activity_info, activing_info_missing, periods
}
\description{
Extract Details From the Bank Account JSON
}
