% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{get_analytics}
\alias{get_analytics}
\title{Get Data From The Analytics API}
\usage{
get_analytics(user, password, book_pk)
}
\arguments{
\item{user}{username for the api}

\item{password}{password for the website}

\item{book_pk}{Book PK}
}
\value{
response_json from api
}
\description{
This function calls the https://api.ocrolus.com/v1/book/summary API.  book_pk is required and is used to return the book data
}
