% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_data.R
\name{get_and_load_book_data}
\alias{get_and_load_book_data}
\title{Load All Data}
\usage{
get_and_load_book_data(
  completed_book_df,
  db_un,
  db_pw,
  db_driver,
  db_server,
  db_name,
  db_port,
  db_schema,
  api_un,
  api_pw,
  slice_amount = 1000,
  core_info_table_name = "OcrolusBooksCoreInfo",
  activity_info_active_table_name = "OcrolusBooksActivityInfoActive",
  activity_info_missing_table_name = "OcrolusBooksActivityInfoMissing",
  periods_table_name = "OcrolusBooksPeriods",
  txns_table_name = "OcrolusBooksTransactions",
  interbank_table_name = "OcrolusAnalyticsInterbankTxns",
  alt_lender_table_name = "OcrolusAnalyticsAlternativeLenderTxs",
  round_number_txs_table_name = "OcrolusAnalyticsRoundNumberTxns",
  ppp_loan_txn_table_name = "OcrolusAnalyticsPPPLoanTxns",
  nsf_txns_table_name = "OcrolusAnalyticsNSFTxns",
  period_balance_mismatches_table_name = "OcrolusAnalyticsPeriodBalanceMismatches",
  outside_source_deposits_table_name = "OcrolusAnalyticsOutsideSourceDeposits",
  withdrawals_max_by_month_table_name = "OcrolusAnalyticsWithdrawalsMaxByMonth",
  deposit_min_by_month_table_name = "OcrolusAnalyticsDepositMinByMonth",
  deposit_max_by_month_table_name = "OcrolusAnalyticsDepositMaxByMonth",
  negative_balances_by_month_table_name = "OcrolusAnalyticsNegativeBalancesByMonth",
  withdrawals_sum_by_month_table_name = "OcrolusAnalyticsWithdrawalsSumByMonth",
  daily_cash_flows_table_name = "OcrolusAnalyticsDailyCashFlows",
  deposits_sum_by_month_table_name = "OcrolusAnalyticsDepositsByMonth",
  estimated_revenue_by_month_table_name = "OcrolusAnalyticsEstimatedRevenueByMonth",
  daily_balances_table_name = "OcrolusAnalyticsDailyBalances",
  completed_table_name = "OcrolusLoadedIDs"
)
}
\arguments{
\item{user}{username for the api}

\item{password}{password for the website}

\item{string}{Fragment of Book name to search for}
}
\value{

}
\description{
This function is designed to be called with the list of data you need to load into the data, this will pull the releveant data, check if it's been loaded
then load it to the DB, There are lots of parameters that you need to pass to make sure the tables are named correctly.
search for a specific book if passed
}
