% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_splicing.R
\name{get_bank_account_from_analyics}
\alias{get_bank_account_from_analyics}
\title{Get Specific Analytics Reponse from API}
\usage{
get_bank_account_from_analyics(analyics_repsonse)
}
\arguments{
\item{analyics_repsonse}{Reponse from the analytics API}
}
\value{
Bank Accounts reponse from the analytics API
}
\description{
Get Specific Analytics Reponse from API
}
