% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{get_book_status}
\alias{get_book_status}
\title{Get Data From the Book Status API}
\usage{
get_book_status(user, password, pk_string)
}
\arguments{
\item{user}{username for the api}

\item{password}{password for the website}

\item{pk_string}{Book PK}
}
\value{
response_json from api
}
\description{
This function calls the https://api.ocrolus.com/v1/book/status API.  string is optional and will allow you to
search for a specific book if the pk_string is passed
}
