% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{get_books}
\alias{get_books}
\title{Get Data From the Book API}
\usage{
get_books(user, password, string = NULL)
}
\arguments{
\item{user}{username for the api}

\item{password}{password for the website}

\item{string}{Fragment of Book name to search for}
}
\value{

}
\description{
This function calls the api.ocrolus.com/v1/books API.  string is optional and will allow you to
search for a specific book if passed
}
