% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_outputs.R
\name{get_docs_df}
\alias{get_docs_df}
\title{Create a DF from the Book Status data returned from the API}
\usage{
get_docs_df(response_json)
}
\arguments{
\item{response_json}{the response json from the book_status api call}
}
\value{
df with cleaned names and renamed
}
\description{
Create a DF from the Book Status data returned from the API
}
