% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_splicing.R
\name{get_min_max_by_month}
\alias{get_min_max_by_month}
\title{used to grab the min and max columns from the bank accounts}
\usage{
get_min_max_by_month(df, col_name)
}
\arguments{
\item{df}{df with the columns to grab}

\item{col_name}{the column containing the min and max values}
}
\value{
df with the selected columns
}
\description{
used to grab the min and max columns from the bank accounts
}
