% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_splicing.R
\name{get_negative_balance}
\alias{get_negative_balance}
\title{Get Negative Value Dates}
\usage{
get_negative_balance(df, col_name)
}
\arguments{
\item{df}{df with the columns to grab}

\item{col}{the column to extract}
}
\value{
df with the selected columns
}
\description{
Get Negative Value Dates
}
