% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_backend.R
\name{get_sql_connection}
\alias{get_sql_connection}
\title{Create Database Backend
Initalize a database backend}
\usage{
get_sql_connection(user_name, password, driver_name, server, database, port)
}
\arguments{
\item{user_name}{username for the api}

\item{password}{password for the website}

\item{driver_name}{Name of the driver being used to ODBC to the DB}

\item{server}{Server IP or Name}

\item{database}{Database Name}

\item{port}{Port of the DB}
}
\value{
connection object
}
\description{
Create Database Backend
Initalize a database backend
}
