% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{get_transactions}
\alias{get_transactions}
\title{Get Data From Transaction API}
\usage{
get_transactions(user, password, book_pk)
}
\arguments{
\item{user}{username for the api}

\item{password}{password for the website}

\item{book_pk}{Book PK}
}
\value{
response_json from api
}
\description{
This function calls thehttps://api.ocrolus.com/v1/transaction API.  book_pk is required
}
