% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_splicing.R
\name{get_txn_info_single_row}
\alias{get_txn_info_single_row}
\title{Get Transactions That return a single value}
\usage{
get_txn_info_single_row(df, col_name)
}
\arguments{
\item{df}{df with the columns to grab}

\item{col_name}{the column containing the single values}
}
\value{
df with the selected columns
}
\description{
Get Transactions That return a single value
}
