% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_backend.R
\name{load_book_information}
\alias{load_book_information}
\title{Load Info to DB
Generic loader function for the analytics part of the Ocrolus API.  This will take a string of key columns to use a check and use that to validate there are not duplicates}
\usage{
load_book_information(
  data_df,
  connection,
  table_name,
  db_schema,
  key_cols,
  analytics = F
)
}
\arguments{
\item{connection}{connection object initalized from get_sql_connection}

\item{table_name}{the name of the table you want to load the book data into}

\item{db_schema}{the schema where you want/have the table}

\item{key_cols}{key columns for the section}

\item{book_df}{df containing the book information, get_books function will give you this data}
}
\description{
Load Info to DB
Generic loader function for the analytics part of the Ocrolus API.  This will take a string of key columns to use a check and use that to validate there are not duplicates
}
