% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_outputs.R
\name{make_bbox_string}
\alias{make_bbox_string}
\title{Make bbox a string value}
\usage{
make_bbox_string(df)
}
\arguments{
\item{df}{DF containing the bbox column}
}
\value{
df with bbox as a string
}
\description{
Make bbox a string value
}
