% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_outputs.R
\name{pluck_map}
\alias{pluck_map}
\title{Internal function used to pluck data from a list column DF}
\usage{
pluck_map(x, name)
}
\arguments{
\item{x}{list data}

\item{name}{name you are trying to pluck}
}
\value{
dataset
}
\description{
Internal function used to pluck data from a list column DF
}
