% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_outputs.R
\name{pluck_map_df}
\alias{pluck_map_df}
\title{Pluck In the Mapping Function}
\usage{
pluck_map_df(x, name)
}
\arguments{
\item{x}{is value to map}

\item{name}{is the value to pluck}
}
\value{
value plucked from list or NA if it wasn't found
}
\description{
Pluck In the Mapping Function
}
