% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_splicing.R
\name{unnest_analytics_df}
\alias{unnest_analytics_df}
\title{Unnest the Analytics DFs}
\usage{
unnest_analytics_df(df, col)
}
\arguments{
\item{df}{df with the columns to grab}

\item{col}{the column to extract}
}
\value{
df with the selected columns
}
\description{
Unnest the Analytics DFs
}
