% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_splicing.R
\name{unnest_bank_transactions}
\alias{unnest_bank_transactions}
\title{Unnest Bank Transactions}
\usage{
unnest_bank_transactions(df, col)
}
\arguments{
\item{analyics_repsonse}{Reponse from the analytics API}
}
\value{
Bank Accounts reponse from the analytics API
}
\description{
Unnest Bank Transactions
}
