% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_backend.R
\name{get_all_data_from_table}
\alias{get_all_data_from_table}
\title{Create Database Backend
Initalize a database backend}
\usage{
get_all_data_from_table(connection, table_name, schema)
}
\arguments{
\item{connection}{connection object}

\item{table_name}{the name of the table to get data from}

\item{schema}{the schema the table is located in}
}
\value{
connection object
}
\description{
Create Database Backend
Initalize a database backend
}
